\name{template-package}
\alias{template-package}
\alias{template}
\docType{package}
\title{
\packageTitle{template}
}
\description{
\packageDescription{template}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{template}
\packageIndices{template}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{template}

Maintainer: \packageMaintainer{template}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
